+++ Runtime/Scripts/Core/Interfaces/IMonitoringTicker.cs
@@ -6,14 +6,11 @@ namespace Baracuda.Monitoring.Interfaces
 {
     internal interface IMonitoringTicker : IMonitoringSubsystem<IMonitoringTicker>
     {
-        /// <summary>
-        /// Toggle validation tick.
-        /// </summary>
         bool ValidationTickEnabled { get; set; }
-        
+
         void AddUpdateTicker(IMonitorUnit unit);
         void RemoveUpdateTicker(IMonitorUnit unit);
-        
+
         void AddValidationTicker(Action tickAction);
         void RemoveValidationTicker(Action tickAction);
     }